SSH - Server Monitoring (#29)
SQUASHING and merging because a commit contained sensitive information. 

* Added ssh_utils.py

Need to call it from experiment.py.
Need to test that it works on Azure.

* Added call to ssh_utils - Needs testing

* Added ability to end server monitoring.

Must start server monitoring, and then pass ssh connection
to end_server_monitoring.

* Added server monitoring to exeriment.py.

* Removed extra call to start server monitoring.

* Fixed root path for project

* ssh_utils now should restart servers on ServerVM

TODO: test this

* Added SSH utility to requirements.txt

* Added Logging to SSH, handle a few errors

Automatically retry SSH connection for a few errors.

* Fixed ssh_utils.

Previously failed because upon every one of Paramiko's
`exec_command`s, it would start a new shell! So if we want to
`cd` or anything, we have to add the multiple commands to one
string.

* Added custom latency, loss, bandwidth

* Addressing linter complaint

* fixing file name

* Modified script to only start server monitoring if on server

* Added documentation

* Renamed ips.json to ips.example.json

* added ips.json to gitignore

* fixed bugs in checking if we're connecting to server VM

Co-authored-by: Jacob Boerma <jboerma@g.hmc.edu>
Co-authored-by: nefrathenrici <nat.henrici@gmail.com>
Co-authored-by: Ubuntu <clinic@Client-VM.arwmfc444k0e3ko1nhv1gjrbff.xx.internal.cloudapp.net>