Making units explicit
According to the `tc` [man page](https://man7.org/linux/man-pages/man8/tc.8.html), a unit-less number represents number of bytes if it's a size, or number of bits per second if it's a rate. I have no idea why it's so inconsistent. 
The `tc-tbf` [man page](https://man7.org/linux/man-pages/man8/tc-tbf.8.html) tells us that `limit` and `burst` are in bytes, as we expect from the `tc` description. Thus I think it's fair to say that our `limit` and `burst` are in bytes. This addition makes it explicit.